# apps/engine/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.AuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 86400
        type: integer
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  handlers.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: password123
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    required:
    - email
    - name
    type: object
  types.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  types.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/types.APIError'
      meta:
        $ref: '#/definitions/types.Meta'
      success:
        type: boolean
    type: object
  types.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      request_id:
        type: string
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@iacstudio.io
    name: IaC Studio Support
    url: https://iacstudio.io/support
  description: Infrastructure as Code management platform with AI-powered recommendations
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://iacstudio.io/terms
  title: IaC Studio API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user (client-side token removal)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.APIResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh JWT token (not implemented yet)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.APIResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/types.APIResponse'
      security:
      - BearerAuth: []
      summary: Refresh token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      summary: Register a new user
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"